+++ test/Volr/Generate/FutharkSpec.hs
@@ -21,7 +21,7 @@ spec = do
     it "can generate a simple NN program" $ do
       let o = compile defaultFutharkProgram (TmNet 1 2)
       let simpleProgram = [r|import "lib/github.com/HnimNart/deeplearning/deep_learning"
-module dl = deep_learning f32
+module dl = deep_learning f64
 let x0 = dl.layers.dense (1, 2) dl.nn.relu 1
 
 let nn = x0
